+++ app/build.gradle
@@ -11,8 +11,8 @@ android {
         applicationId "com.app.showtime"
         minSdk 26
         targetSdk 32
-        versionCode 1
-        versionName "1.0"
+        versionCode 5
+        versionName "5"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -53,7 +53,7 @@ dependencies {
     implementation ('io.socket:socket.io-client:2.0.0') {
         exclude group: 'org.json', module: 'json'
     }
-
+    implementation 'com.google.android.gms:play-services-auth:20.5.0'
 }
